backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
Pillow
numpy
pandas
scikit-learn
tensorflow>=2.20.0
joblib
gunicorn
celery
redis
dj-database-url